+++ Unofficial.SignalR.Protobuf/MessageSerializers/StreamInvocationMessageSerializer.cs
@@ -13,7 +13,8 @@ namespace Unofficial.SignalR.Protobuf.MessageSerializers
         public ProtobufMessageType EnumType => ProtobufMessageType.StreamInvocation;
         public Type MessageType => typeof(StreamInvocationMessage);
 
-        public void WriteMessage(HubMessage message, IBufferWriter<byte> output, IReadOnlyDictionary<Type, int> protobufTypeToIndexMap)
+        public void WriteMessage(HubMessage message, IBufferWriter<byte> output,
+            IReadOnlyDictionary<Type, short> protobufTypeToIndexMap)
         {
             var streamInvocationMessage = (StreamInvocationMessage) message;
